Minor modifications to reflect changes in DDF-2571
