+++ catalog/async/catalog-async-data/pom.xml
@@ -24,7 +24,7 @@
     <modelVersion>4.0.0</modelVersion>
     <artifactId>catalog-async-data</artifactId>
     <name>Catalog :: Async :: Data :: Impl</name>
-    <packaging>bundle</packaging>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
@@ -52,23 +52,34 @@
 
     <build>
         <plugins>
+            <!--<plugin>-->
+                <!--<groupId>org.apache.felix</groupId>-->
+                <!--<artifactId>maven-bundle-plugin</artifactId>-->
+                <!--<extensions>true</extensions>-->
+                <!--<configuration>-->
+                    <!--<instructions>-->
+                        <!--<Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>-->
+                        <!--<Bundle-Name>${project.name}</Bundle-Name>-->
+                        <!--<Embed-Dependency>-->
+                            <!--catalog-core-api-impl,-->
+                            <!--platform-util-->
+                        <!--</Embed-Dependency>-->
+                        <!--<Export-Package>-->
+                            <!--org.codice.ddf.catalog.async.data*-->
+                        <!--</Export-Package>-->
+                    <!--</instructions>-->
+                <!--</configuration>-->
+            <!--</plugin>-->
             <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <extensions>true</extensions>
-                <configuration>
-                    <instructions>
-                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Name>${project.name}</Bundle-Name>
-                        <Embed-Dependency>
-                            catalog-core-api-impl,
-                            platform-util
-                        </Embed-Dependency>
-                        <Export-Package>
-                            org.codice.ddf.catalog.async.data*
-                        </Export-Package>
-                    </instructions>
-                </configuration>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
             <plugin>
                 <groupId>org.jacoco</groupId>
