+++ catalog/async/catalog-async-data/src/main/java/org/codice/ddf/catalog/async/data/impl/ProcessDeleteItemImpl.java
@@ -11,29 +11,22 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.async.data;
+package org.codice.ddf.catalog.async.data.impl;
 
-import java.io.Serializable;
-import java.util.List;
-import java.util.Map;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessDeleteItem;
 
-import org.codice.ddf.catalog.async.data.api.internal.ProcessItem;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessRequest;
+import ddf.catalog.data.Metacard;
 
-import ddf.catalog.operation.impl.OperationImpl;
+public class ProcessDeleteItemImpl implements ProcessDeleteItem {
 
-public class ProcessRequestImpl<T extends ProcessItem> extends OperationImpl
-        implements ProcessRequest {
+    private final Metacard metacard;
 
-    private List<T> processItems;
-
-    public ProcessRequestImpl(List<T> processItems, Map<String, Serializable> properties) {
-        super(properties);
-        this.processItems = processItems;
+    public ProcessDeleteItemImpl(Metacard metacard) {
+        this.metacard = metacard;
     }
 
     @Override
-    public List<T> getProcessItems() {
-        return processItems;
+    public Metacard getMetacard() {
+        return metacard;
     }
 }
