+++ catalog/async/catalog-async-postingestplugin/src/main/java/org/codice/ddf/catalog/async/processingplugin/ProcessingPostIngestPlugin.java
@@ -23,16 +23,16 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.commons.collections.CollectionUtils;
-import org.codice.ddf.catalog.async.data.ProcessCreateItemImpl;
-import org.codice.ddf.catalog.async.data.ProcessDeleteItemImpl;
-import org.codice.ddf.catalog.async.data.ProcessRequestImpl;
-import org.codice.ddf.catalog.async.data.ProcessResourceImpl;
-import org.codice.ddf.catalog.async.data.ProcessUpdateItemImpl;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessCreateItem;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessDeleteItem;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessRequest;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessResource;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessUpdateItem;
+import org.codice.ddf.catalog.async.data.impl.ProcessCreateItemImpl;
+import org.codice.ddf.catalog.async.data.impl.ProcessDeleteItemImpl;
+import org.codice.ddf.catalog.async.data.impl.ProcessRequestImpl;
+import org.codice.ddf.catalog.async.data.impl.ProcessResourceImpl;
+import org.codice.ddf.catalog.async.data.impl.ProcessUpdateItemImpl;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessCreateItem;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessDeleteItem;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessRequest;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessResource;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessUpdateItem;
 import org.codice.ddf.catalog.async.processingframework.api.internal.ProcessingFramework;
 import org.codice.ddf.security.common.Security;
 import org.slf4j.Logger;
