+++ catalog/async/catalog-async-processingframework-api/src/main/java/org/codice/ddf/catalog/async/processingframework/api/internal/ProcessingFramework.java
@@ -13,10 +13,10 @@
  */
 package org.codice.ddf.catalog.async.processingframework.api.internal;
 
-import org.codice.ddf.catalog.async.data.api.internal.ProcessCreateItem;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessDeleteItem;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessRequest;
-import org.codice.ddf.catalog.async.data.api.internal.ProcessUpdateItem;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessCreateItem;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessDeleteItem;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessRequest;
+import org.codice.ddf.catalog.async.data.impl.api.internal.ProcessUpdateItem;
 
 /**
  * <p>
