+++ catalog/catalog-app/src/main/resources/features.xml
@@ -443,7 +443,7 @@
         <bundle>mvn:org.apache.mina/mina-core/${mina.version}</bundle>
     </feature>
 
-    <feature name="catalog-async-inmemory" install="auto" version="${project.version}">
+    <feature name="catalog-async-inmemory" install="manual" version="${project.version}">
         <feature prerequisite="true">catalog-app</feature>
         <bundle>mvn:ddf.catalog.async/catalog-async-data-api/${project.version}</bundle>
         <bundle>mvn:ddf.catalog.async/catalog-async-plugin-api/${project.version}</bundle>
